cheri: Add uinptr_t to unsigned long casts in net/
Add correct casts between uintptr_t and unsigned long to the
net/ subsystem.

+++ net/ipv4/icmp.c
@@ -533,7 +533,7 @@ static struct rtable *icmp_route_lookup(struct net *net,
 			err = PTR_ERR(rt2);
 	} else {
 		struct flowi4 fl4_2 = {};
-		unsigned long orefdst;
+		uintptr_t orefdst;
 
 		fl4_2.daddr = fl4_dec.saddr;
 		rt2 = ip_route_output_key(net, &fl4_2);
