+++ net/ipv4/ip_fragment.c
@@ -373,7 +373,7 @@ static int ip_frag_queue(struct ipq *qp, struct sk_buff *skb)
 
 	if (qp->q.flags == (INET_FRAG_FIRST_IN | INET_FRAG_LAST_IN) &&
 	    qp->q.meat == qp->q.len) {
-		unsigned long orefdst = skb->_skb_refdst;
+		uintptr_t orefdst = skb->_skb_refdst;
 
 		skb->_skb_refdst = 0UL;
 		err = ip_frag_reasm(qp, skb, prev_tail, dev);
